requirements.txt
beautifulsoup4==4.6.0
billboard.py==4.0.0
certifi==2017.11.5
chardet==3.0.4
idna==2.6
lyricwikia==0.1.4
requests==2.18.4
urllib3==1.22
wheel==0.24.0
